refactor: codingzone 도메인 관련 VO 파일 이동
