+++ backend/src/main/java/com/icehufs/icebreaker/domain/article/domain/vo/CommentListItem.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.object;
+package com.icehufs.icebreaker.domain.article.domain.vo;
 
 import java.util.ArrayList;
 import java.util.List;
